==> 03.01 Loading csv files into dataframes.ipynb <==
data_path = '/Users/quangly/github/Spark SQL/Data'

file_path = data_path + "/location_temp.csv"
df1 = spark.read.format("csv").option("header", "true").load(file_path)

df1.head(10)

df1.show()

df1.count()

file_path_no_header = data_path + "/utilization.csv"
df2 = spark.read.format("csv").option("header", "false").option("inferSchema","true").load(file_path_no_header)

df2.count()

df2.head(10)

df2 = df2.withColumnRenamed("_c0", "event_datetime") \
        .withColumnRenamed ("_c1", "server_id")       \
        .withColumnRenamed("_c2", "cpu_utilization")  \
        .withColumnRenamed("_c3", "free_memory")      \
        .withColumnRenamed("_c4", "session_count")

df2.head(10)

df2.show()

json_df1_path = data_path + "/location_temp.json"
df1.write.json(json_df1_path)

json_df2_path = data_path + "/utilization.json"
df2.write.json(json_df2_path)
