==> 03.03 Basic Dataframe Operations.ipynb <==
from pyspark.sql import SparkSession

data_path = '/Users/quangly/github/Spark SQL/Data'

json_df2_path = data_path + "/utilization.json"
df2 = spark.read.format("json").load(json_df2_path)

df2.head(20)

df2.show(10)

df2.count()

df2.columns

df2.describe()

df2_sample = df2.sample(False,fraction=0.1)

df2_sample.show()

df2_sort = df2_sample.sort('event_datetime',ascending=False)

df2_sort.show()

df2.describe().show()

df
