==> 03.05 Aggregating using Dataframe API.ipynb <==
from pyspark.sql import SparkSession

data_path = '/Users/quangly/github/Spark SQL/Data'

file_path = data_path + "/location_temp.csv"
df1 = spark.read.format("csv").option("header", "true").load(file_path)

df1.show(10)

df1.groupBy("location_id").count().show()

df1.orderBy("location_id").show(10)

Average Temperature at Each Location

df1.groupby('location_id').agg({'temp_celcius': 'mean'}).orderBy("location_id").show()

df1.groupby('location_id').agg({'temp_celcius': 'max'}).orderBy("location_id").show()
