==> 03.06 Sampling using Dataframe API.ipynb <==
from pyspark.sql import SparkSession
data_path = '/Users/quangly/github/Spark SQL/Data'

file_path = data_path + "/location_temp.csv"
df1 = spark.read.format("csv").option("header", "true").load(file_path)

df1.show(10)

df1.count()

withReplacement - when removing a row never put back in to prevent duplicates

df_s1 = df1.sample(fraction=0.1,withReplacement=False)

df_s1.count()

df_s1.groupBy("location_id").agg({'temp_celcius': 'mean'}).orderBy("location_id").show(10)

df1.groupBy("location_id").agg({'temp_celcius': 'mean'}).orderBy("location_id").show(10)

df_s2 = df1.sample(fraction=0.001,withReplacement=False)

df_s2.groupBy("location_id").agg({'temp_celcius': 'mean'}).orderBy("location_id").show(10)
