==> 03.07 Saving Data from Dataframes.ipynb <==
from pyspark.sql import SparkSession
data_path = '/Users/quangly/github/Spark SQL/Data'

file_path = data_path + "/location_temp.csv"
df1 = spark.read.format("csv").option("header", "true").load(file_path)

df1.show(10)

df1.write.mode('overwrite').csv('df1.csv')

#! everything after in shell is a command
! ls df1.csv

! head df1.csv/part-00000-5ab67295-1daa-4529-bde9-967d8f87e6b0-c000.csv

df1.write.mode('overwrite').json('df1.json')

! ls df1.json

! head df1.json/part-00000-6c2d74cb-007b-49b7-b8b2-361e9cc1a078-c000.json

 df1.write.mode('overwrite').parquet('df1.parquet')

! ls head df1.parquet

! head df1.parquet/part-00000-0d3e8e98-8f5d-43f5-a907-d3270f1a2b24-c000.snappy.parquet



df1.write.mode('overwrite').csv('df1-gzip.csv', compression="gzip")  # Python-only
