==> 04.02 Filtering Dataframes with SQL.ipynb <==
from pyspark.sql import SparkSession
spark = SparkSession \
    .builder \
    .appName("Spark SQL Query Dataframes") \
    .getOrCreate()
data_path = '/Users/quangly/github/Spark SQL/Data'

json_df2_path = data_path + "/utilization.json"
df = spark.read.format("json").load(json_df2_path)

df.show(10)

df.createOrReplaceTempView("utilization")

df_sql = spark.sql("SELECT * FROM utilization WHERE server_id = 120")
df_sql.show()

df_sql.count()

df_sql = spark.sql("SELECT server_id, session_count FROM utilization WHERE session_count > 70")
df_sql.show()

df_sql.count()

df_sql = spark.sql("SELECT server_id, session_count FROM utilization WHERE session_count > 70 AND server_id = 120")
df_sql.show()

df_sql.count()

df_sql = spark.sql("SELECT server_id, session_count \
                   FROM utilization \
                   WHERE session_count > 70 AND server_id = 120 \
                   ORDER BY session_count DESC")
df_sql.show()
