==> 04.03 Aggregating Dataframes with SQL.ipynb <==
from pyspark.sql import SparkSession

from pyspark.sql import SparkSession
spark = SparkSession \
    .builder \
    .appName("Spark SQL Query Dataframes") \
    .getOrCreate()
data_path = '/Users/quangly/github/Spark SQL/Data'

json_df2_path = data_path + '/utilization.json'
df = spark.read.format("json").load(json_df2_path)

df.show(10)

df.createOrReplaceTempView("utilization")

df_count = spark.sql("SELECT count(*) FROM utilization")
df_count.show()

df_sql = spark.sql("SELECT count(*) \
                    FROM utilization \
                    WHERE session_count > 70")
df_sql.show()

df_sql = spark.sql("SELECT server_id, count(*) \
                    FROM utilization \
                    WHERE session_count > 70 \
                    GROUP BY server_id")
df_sql.show()

df_sql = spark.sql("SELECT server_id, count(*) \
                    FROM utilization \
                    WHERE session_count > 70 \
                    GROUP BY server_id \
                    ORDER BY count(*) DESC")
df_sql.show()

df_sql = spark.sql("SELECT server_id, min(session_count), avg(session_count), max(session_count) \
                    FROM utilization \
                    WHERE session_count > 70 \
                    GROUP BY server_id \
                    ORDER BY count(*) DESC")
df_sql.show()

df_sql = spark.sql("SELECT server_id, count(*), min(session_count), round(avg(session_count),2), max(session_count) \
                    FROM utilization \
                    WHERE session_count > 70 \
                    GROUP BY server_id \
                    ORDER BY count(*) DESC")
df_sql.show()
