==> 04.04 Joining Dataframes with SQL.ipynb <==
from pyspark.sql import SparkSession
from pyspark.sql.types import IntegerType
spark = SparkSession \
    .builder \
    .appName("Spark SQL Query Dataframes") \
    .getOrCreate()
data_path = '/Users/quangly/github/Spark SQL/Data'

json_df_path = data_path + "/utilization.json"
df_util = spark.read.format("json").load(json_df_path)

df_util.show(10)

df_util.createOrReplaceTempView("utilization")

csv_df_path = data_path + "/server_name.csv"
df_server = spark.read.csv(csv_df_path,header=True)

df_server.show()

df_server.createOrReplaceTempView("server_name")

df_count = spark.sql("SELECT DISTINCT server_id FROM utilization ORDER BY server_id")
df_count.show()

spark.sql("SELECT min(server_id), max(server_id) FROM utilization").show()

spark.sql("SELECT * FROM server_name").show()

df_join = spark.sql("SELECT u.server_id, sn.server_name, u.session_count \
                     FROM utilization u \
                     INNER JOIN server_name sn \
                     ON sn.server_id = u.server_id")
df_join.show()
