==> 04.05 De-duplicating.ipynb <==
from pyspark.sql import SparkSession
from pyspark.sql import Row

spark = SparkSession \
    .builder \
    .appName("Spark SQL Query Dataframes") \
    .getOrCreate()
data_path = '/Users/quangly/github/Spark SQL/Data'

sc is global variable SparkContext
parallelize - to create a dataframe. parallelize data structure

df_dup = sc.parallelize([Row(server_name='101 Server', cpu_utilization=85, session_count=80), \
                             Row(server_name='101 Server', cpu_utilization=80, session_count=90),
                             Row(server_name='102 Server', cpu_utilization=85, session_count=80),
                             Row(server_name='102 Server', cpu_utilization=85, session_count=80)]).toDF()

df_dup.show()

df_dup.drop_duplicates().show()

df_dup.drop_duplicates(['server_name']).show()

df_dup.show()
