==> 04.06 Working with NAs - null values in DataFrames.ipynb <==
from pyspark.sql import SparkSession
from pyspark.sql import Row
from pyspark.sql.functions import lit
#lit which will allow to create a literal column
from pyspark.sql.types import StringType

df = sc.parallelize([Row(server_name='101 Server', cpu_utilization=85, session_count=80), \
                             Row(server_name='101 Server', cpu_utilization=80, session_count=90),
                             Row(server_name='102 Server', cpu_utilization=85, session_count=40),
                             Row(server_name='103 Server', cpu_utilization=70, session_count=80),
                             Row(server_name='104 Server', cpu_utilization=60, session_count=80)]).toDF()

df.show()

# add column na_col with literal value of None/Null and cast as String Type

df_na = df.withColumn('na_col', lit(None).cast(StringType()))

df_na.show()

df_na.fillna('A').show()

df2 = df_na.fillna('A').union(df_na)

df2.show()

df2.na.drop().show()

df2.createOrReplaceTempView("na_table")

spark.sql("SELECT * FROM na_table").show()

spark.sql("SELECT * FROM na_table WHERE na_col IS NULL").show()

spark.sql("SELECT * FROM na_table WHERE na_col IS NOT NULL").show()
