==> 05.01 Exploratory Analysis.ipynb <==
from pyspark.sql import SparkSession

spark = SparkSession.builder.getOrCreate()

data_path = '/Users/quangly/github/Spark SQL/Data'

json_df2_path = data_path + "/utilization.json"
df_util = spark.read.format("json").load(json_df2_path)

df_util.show(10)

df_util.createOrReplaceTempView("utilization")

df_util.count()

#describe() produces another df with summary

df_util.describe().show()

df_util.describe('cpu_utilization','free_memory').show()

#correlation
# no correlation : 0
# if one value goes up other value always go up: 1
# if one value goes up other value goes down: negative

df_util.stat.corr('cpu_utilization','free_memory')

df_util.stat.corr('session_count','free_memory')

# freqItems - what are freq occuring items

df_util.stat.freqItems(('server_id','session_count')).show()

# withReplacement = False - as you choose, something out of population, do you want to put it back into population which could create dups

df_util_sample = df_util.sample(fraction=0.05, withReplacement=False)
df_util_sample.count()

spark.sql('SELECT min(cpu_utilization), max(cpu_utilization), stddev(cpu_utilization) FROM utilization').show()

spark.sql('SELECT server_id, min(cpu_utilization), max(cpu_utilization), stddev(cpu_utilization) \
           FROM utilization \
           GROUP BY server_id').show()

#buckets or histograms in ranges.

spark.sql('SELECT server_id, FLOOR(cpu_utilization*100/10) bucket FROM utilization').show()

spark.sql('SELECT count(*), FLOOR(cpu_utilization*100/10) bucket FROM utilization GROUP BY bucket ORDER BY bucket').show()
