==> 05.02 Timeseries Analysis .ipynb <==
from pyspark.sql import SparkSession

spark = SparkSession.builder.getOrCreate()

data_path = '/Users/quangly/github/Spark SQL/Data'

json_df2_path = data_path + "/utilization.json"
df_util = spark.read.format("json").load(json_df2_path)

df_util.createOrReplaceTempView("utilization")

spark.sql('SELECT server_id, min(cpu_utilization), max(cpu_utilization), stddev(cpu_utilization) \
           FROM utilization \
           GROUP BY server_id').show()

sql_window = "SELECT event_datetime, server_id, cpu_utilization,  \
         avg(cpu_utilization) OVER (PARTITION BY server_id) avg_server_util \
FROM  \
      utilization"

spark.sql(sql_window).show()

# sometimes want to compare a value such as current cpu of a server with average of that server. not entire population
# windowing function written with PARTITION BY

sql_window2 = "SELECT event_datetime, server_id, cpu_utilization,  \
         avg(cpu_utilization) OVER (PARTITION BY server_id) avg_server_util, \
         cpu_utilization - avg(cpu_utilization) OVER (PARTITION BY server_id) delta_server_util \
         FROM utilization"

spark.sql(sql_window2).show()

# calculate neighborhood of a row
# sliding window - ex. last 3 values and average them

sql_window3 = "SELECT event_datetime, server_id, cpu_utilization,  \
                      avg(cpu_utilization) OVER (PARTITION BY server_id ORDER BY event_datetime \
                                    ROWS BETWEEN 1 PRECEDING AND 1 FOLLOWING) avg_server_util \
                FROM  \
                      utilization"


spark.sql(sql_window3).show()

(0.68+0.58)/2

(0.68+0.58+0.55)/3
