==> 05.03b Machine Learning - Clustering-v2.ipynb <==
from pyspark.sql import SparkSession
from pyspark.ml.linalg import Vectors
from pyspark.ml.feature import VectorAssembler
from pyspark.ml.clustering import KMeans

spark = SparkSession.builder.getOrCreate()

data_path = '/Users/quangly/github/Spark SQL/Data'

# exploratory data analysis, Clustering. natural groupings
#divide data set into 3 logical datasets 

csv_path = data_path + "/clustering_dataset.csv"
cluster_df = spark.read.format("csv") \
                .option("header", True)\
                .option("inferSchema", True) \
                .load(csv_path)

cluster_df.show(75)

# group based on cpu, free memory, session count
# ml libraries work with vectors (like an array)
# create vector to store all 3 values called "features"

vectorAssembler = VectorAssembler(inputCols=["col1", "col2", "col3"], outputCol="features")

# vectorized cluster data frame.

vcluster_df = vectorAssembler.transform(cluster_df)

vcluster_df.show()

# KMeans ML algorithm
# K = number of clusters
# takes a seed
# kmeans a structure to run kmeans algorithm

kmeans = KMeans().setK(3)
kmeans = kmeans.setSeed(1)

# fit - take input data and apply algorithm

kmodel = kmeans.fit(vcluster_df)

# Critical thing of KMeans model is Cluster Center
# we have set of 3 centers, each center is represented by 3 values similar to x, y z in space \ 
# but cpu utilization, free memory, and sessioncount.
# each of these values indicate the center of these 3 values.
# all of our rows in our dataframe fit or fall into one of these 3 clusters.
# we can measure by the distance of the feature vector of each of the center

kmodel.clusterCenters()
