==> 05.03f Machine Learning - Classification-IRIS Decision Trees.ipynb <==
# https://scikit-learn.org/stable/modules/tree.html
# Decision Trees (DTs) are a non-parametric supervised learning method used for classification and regression. The goal is to create a model that predicts the value of a target variable by learning simple decision rules inferred from the data features. A tree can be seen as a piecewise constant approximation.

# For instance, in the example below, decision trees learn from data to approximate a sine curve with a set of if-then-else decision rules. The deeper the tree, the more complex the decision rules and the fitter the model.


from pyspark.sql import SparkSession
from pyspark.sql.functions import *
from pyspark.ml.feature import StringIndexer
from pyspark.ml.linalg import Vectors
from pyspark.ml.feature import VectorAssembler
from pyspark.ml.classification import NaiveBayes
from pyspark.ml.evaluation import MulticlassClassificationEvaluator
from pyspark.ml.classification import MultilayerPerceptronClassifier
from pyspark.ml.classification import DecisionTreeClassifier

spark = SparkSession.builder.getOrCreate()

data_path = '/Users/quangly/github/Spark SQL/Data'

# exploratory data analysis, Clustering. natural groupings
#divide data set into 3 logical datasets 

csv_path = data_path + "/iris.txt"
iris_df = spark.read.format("csv") \
                .option("inferSchema", True) \
                .load(csv_path)

# group based on cpu, free memory, session count
# ml libraries work with vectors (like an array)
# create vector to store all 3 values called "features"

iris_df.show(5)

iris_df = iris_df.select(col("_c0").alias("sepal_length"),
        col("_c1").alias("sepal_width"),
        col("_c2").alias("petal_length"),
        col("_c3").alias("petal_width"),
        col("_c4").alias("species"))


iris_df.take(1)

vectorAssembler = VectorAssembler(inputCols=["sepal_length", "sepal_width", "petal_length", "petal_width"], outputCol="features")

#vectorized iris df

viris_df = vectorAssembler.transform(iris_df)

viris_df.show()

# Part of PreProcessing. Convert label name (species name) into numeric value.
# StringIndexer - 

indexer = StringIndexer(inputCol="species", outputCol="label")
ivirus_df = indexer.fit(viris_df).transform(viris_df)
ivirus_df.show(3)

#Naive Bayes

splits = ivirus_df.randomSplit([0.6,0.4])
train_df = splits[0]
test_df = splits[1]

print(train_df.count())
print(test_df.count())
print(viris_df.count())

nb = NaiveBayes(modelType="multinomial")

nbmodel = nb.fit(train_df)

predictions_df = nbmodel.transform(test_df)

predictions_df.take(1)

# prediction=0.0 is index of the species Setosa which is correct

evaluator = MulticlassClassificationEvaluator(labelCol="label", predictionCol="prediction", metricName="accuracy")

nbaccuracy = evaluator.evaluate(predictions_df)

nbaccuracy

#first layer has same number of nodes as inputs. 4 measures so 4. last layer has same number of types of output 3
# layers in between so the multilayer perceptron has learn how to classify correctly. 2 rows 5 neurons each
layers = [4,5,5, 3]

mlp = MultilayerPerceptronClassifier(layers = layers, seed=1)

mlp_model = mlp.fit(train_df)

mlp_predictions = mlp_model.transform(test_df)

mlp_evaluator = MulticlassClassificationEvaluator(metricName="accuracy")

mlp_accuracy = mlp_evaluator.evaluate(mlp_predictions)

mlp_accuracy

#higher than Naive Bayes

# Decision Trees

dt = DecisionTreeClassifier(labelCol="label", featuresCol="features")

dt_model = dt.fit(train_df)

dt_predictions = dt_model.transform(test_df)

dt_evaluator = MulticlassClassificationEvaluator(labelCol="label", predictionCol="prediction", metricName="accuracy")

dt_accuracy = dt_evaluator.evaluate(dt_predictions)

dt_accuracy

# Decision Tree and MLP worked well. but Decition trees dont' require us to make any configuration decisions
# Naive Bayes can work well when attributes are indepedent of each other.
