==> 05.04d Machine Learning - Linear Regression-Combined Cycle Power Plant-Gradient Boosted Tree-Regression.ipynb <==
from pyspark.sql import SparkSession
from pyspark.ml.regression import LinearRegression, DecisionTreeRegressor, GBTRegressor
from pyspark.ml.feature import VectorAssembler
from pyspark.ml.evaluation import RegressionEvaluator

spark = SparkSession.builder.getOrCreate()

data_path = '/Users/quangly/github/Spark SQL/Data'

file_path = data_path + "/CCPP/power_plant.csv"
pp_df = spark.read.format("csv") \
    .option("header", True) \
    .option("inferSchema", True) \
    .load(file_path)
pp_df
# PE is Power Generated"

#linear regression is an ML technique to predict 1 variable based on another variable
# for example knowing cpu utilization will allow you to guess number of sessions or free memory

vectorAssembler = VectorAssembler(inputCols=["AT","V","AP","RH"], outputCol="features")
vpp_df = vectorAssembler.transform(pp_df)
vpp_df.show()

# predict session count as label Col.

# lr is our model which is our core structure

splits = vpp_df.randomSplit([.7,.3])

train_df = splits[0]
test_df = splits[1]

train_df.count()

test_df.count()

vpp_df.count()

lr = LinearRegression(featuresCol="features",labelCol="PE")

lr_model = lr.fit(vpp_df)

# a linear regression. represented by coefficient and the intercept. y = mx + b
# slope is m = coefficient

lr_model.coefficients

# where does line intercept y axis intercept

lr_model.intercept

# what is the error for LR model

lr_model.summary.rootMeanSquaredError

error_rate = lr_model.summary.rootMeanSquaredError/lr_model.intercept

error_rate

lr_model.write().overwrite().save("lr1.model")

#DecisionTree Regressor


dt = DecisionTreeRegressor(featuresCol="features", labelCol="PE")

dt_model = dt.fit(train_df)
dt_predictions = dt_model.transform(test_df)

# when working with Classification evaluator, we use "accuracy"
# but when using numeric, we want to see how far off we are so we use error metric - in this case root mean squared error

dt_evaluator = RegressionEvaluator(labelCol="PE", predictionCol="prediction", metricName="rmse")

rmse = dt_evaluator.evaluate(dt_predictions)

rmse

# about 1% of the average power generated (450 - 500)

# Gradient Boosted Trees

gbt = GBTRegressor(featuresCol="features", labelCol="PE")

gbt_model = gbt.fit(train_df)

gbt_predictions = gbt_model.transform(test_df)

gbt_evaluator = RegressionEvaluator(labelCol="PE", predictionCol="prediction", metricName="rmse")
gbt_rmse = gbt_evaluator.evaluate(gbt_predictions)

gbt_rmse

#slightly better than LinearRegression and DecisionTreeRegression
# GradientBoostedTrees is best but takes more time.
